move collectors outside of implicit collection dir
