+++ x-pack/plugins/asset_manager/server/lib/accessors/hosts/get_hosts_by_signals.ts
@@ -7,7 +7,7 @@
 
 import { Asset } from '../../../../common/types_api';
 import { GetHostsOptionsInjected } from '.';
-import { collectHosts } from '../../implicit_collection/collectors/hosts';
+import { collectHosts } from '../../collectors/hosts';
 
 export async function getHostsBySignals(
   options: GetHostsOptionsInjected
