+++ x-pack/plugins/asset_manager/server/lib/collectors/helpers.ts
