+++ x-pack/plugins/asset_manager/server/lib/collectors/hosts.ts
@@ -6,11 +6,11 @@
  */
 
 import { estypes } from '@elastic/elasticsearch';
-import { Asset } from '../../../../common/types_api';
+import { Asset } from '../../../common/types_api';
 import { CollectorOptions, QUERY_MAX_SIZE } from '.';
 import { withSpan } from './helpers';
 
-export async function collectContainers({
+export async function collectHosts({
   client,
   from,
   to,
@@ -20,19 +20,18 @@ export async function collectContainers({
 }: CollectorOptions) {
   const { metrics, logs, traces } = sourceIndices;
   const dsl: estypes.SearchRequest = {
-    index: [traces, logs, metrics],
+    index: [metrics, logs, traces],
     size: QUERY_MAX_SIZE,
-    collapse: {
-      field: 'container.id',
-    },
-    sort: [{ 'container.id': 'asc' }],
+    collapse: { field: 'host.hostname' },
+    sort: [{ 'host.hostname': 'asc' }],
     _source: false,
     fields: [
-      'kubernetes.*',
-      'cloud.provider',
-      'orchestrator.cluster.name',
-      'host.name',
+      '@timestamp',
+      'cloud.*',
+      'container.*',
       'host.hostname',
+      'kubernetes.*',
+      'orchestrator.cluster.name',
     ],
     query: {
       bool: {
@@ -46,11 +45,11 @@ export async function collectContainers({
             },
           },
         ],
+        must: [{ exists: { field: 'host.hostname' } }],
         should: [
-          { exists: { field: 'kubernetes.container.id' } },
-          { exists: { field: 'kubernetes.pod.uid' } },
           { exists: { field: 'kubernetes.node.name' } },
-          { exists: { field: 'host.hostname' } },
+          { exists: { field: 'kubernetes.pod.uid' } },
+          { exists: { field: 'container.id' } },
         ],
       },
     },
@@ -65,25 +64,45 @@ export async function collectContainers({
   const result = withSpan({ transaction, name: 'processing_response' }, async () => {
     const assets = esResponse.hits.hits.reduce<Asset[]>((acc: Asset[], hit: any) => {
       const { fields = {} } = hit;
-      const containerId = fields['container.id'];
-      const podUid = fields['kubernetes.pod.uid'];
-      const nodeName = fields['kubernetes.node.name'];
+      const hostName = fields['host.hostname'];
+      const k8sNode = fields['kubernetes.node.name'];
+      const k8sPod = fields['kubernetes.pod.uid'];
 
-      const parentEan = podUid ? `pod:${podUid}` : `host:${fields['host.hostname']}`;
+      const hostEan = `host:${k8sNode || hostName}`;
 
-      const container: Asset = {
+      const host: Asset = {
         '@timestamp': new Date().toISOString(),
-        'asset.kind': 'container',
-        'asset.id': containerId,
-        'asset.ean': `container:${containerId}`,
-        'asset.parents': [parentEan],
+        'asset.kind': 'host',
+        'asset.id': k8sNode || hostName,
+        'asset.name': k8sNode || hostName,
+        'asset.ean': hostEan,
       };
 
-      if (nodeName) {
-        container['asset.references'] = [`host:${nodeName}`];
+      if (fields['cloud.provider']) {
+        host['cloud.provider'] = fields['cloud.provider'];
       }
 
-      acc.push(container);
+      if (fields['cloud.instance.id']) {
+        host['cloud.instance.id'] = fields['cloud.instance.id'];
+      }
+
+      if (fields['cloud.service.name']) {
+        host['cloud.service.name'] = fields['cloud.service.name'];
+      }
+
+      if (fields['cloud.region']) {
+        host['cloud.region'] = fields['cloud.region'];
+      }
+
+      if (fields['orchestrator.cluster.name']) {
+        host['orchestrator.cluster.name'] = fields['orchestrator.cluster.name'];
+      }
+
+      if (k8sPod) {
+        host['asset.children'] = [`pod:${k8sPod}`];
+      }
+
+      acc.push(host);
 
       return acc;
     }, []);
