+++ x-pack/plugins/asset_manager/server/lib/collectors/pods.ts
@@ -6,11 +6,11 @@
  */
 
 import { estypes } from '@elastic/elasticsearch';
-import { Asset } from '../../../../common/types_api';
+import { Asset } from '../../../common/types_api';
 import { CollectorOptions, QUERY_MAX_SIZE } from '.';
 import { withSpan } from './helpers';
 
-export async function collectHosts({
+export async function collectPods({
   client,
   from,
   to,
@@ -22,16 +22,17 @@ export async function collectHosts({
   const dsl: estypes.SearchRequest = {
     index: [metrics, logs, traces],
     size: QUERY_MAX_SIZE,
-    collapse: { field: 'host.hostname' },
-    sort: [{ 'host.hostname': 'asc' }],
+    collapse: {
+      field: 'kubernetes.pod.uid',
+    },
+    sort: [{ 'kubernetes.pod.uid': 'asc' }],
     _source: false,
     fields: [
-      '@timestamp',
-      'cloud.*',
-      'container.*',
-      'host.hostname',
       'kubernetes.*',
+      'cloud.provider',
       'orchestrator.cluster.name',
+      'host.name',
+      'host.hostname',
     ],
     query: {
       bool: {
@@ -45,11 +46,9 @@ export async function collectHosts({
             },
           },
         ],
-        must: [{ exists: { field: 'host.hostname' } }],
-        should: [
-          { exists: { field: 'kubernetes.node.name' } },
+        must: [
           { exists: { field: 'kubernetes.pod.uid' } },
-          { exists: { field: 'container.id' } },
+          { exists: { field: 'kubernetes.node.name' } },
         ],
       },
     },
@@ -64,45 +63,27 @@ export async function collectHosts({
   const result = withSpan({ transaction, name: 'processing_response' }, async () => {
     const assets = esResponse.hits.hits.reduce<Asset[]>((acc: Asset[], hit: any) => {
       const { fields = {} } = hit;
-      const hostName = fields['host.hostname'];
-      const k8sNode = fields['kubernetes.node.name'];
-      const k8sPod = fields['kubernetes.pod.uid'];
+      const podUid = fields['kubernetes.pod.uid'];
+      const nodeName = fields['kubernetes.node.name'];
+      const clusterName = fields['orchestrator.cluster.name'];
 
-      const hostEan = `host:${k8sNode || hostName}`;
-
-      const host: Asset = {
+      const pod: Asset = {
         '@timestamp': new Date().toISOString(),
-        'asset.kind': 'host',
-        'asset.id': k8sNode || hostName,
-        'asset.name': k8sNode || hostName,
-        'asset.ean': hostEan,
+        'asset.kind': 'pod',
+        'asset.id': podUid,
+        'asset.ean': `pod:${podUid}`,
+        'asset.parents': [`host:${nodeName}`],
       };
 
       if (fields['cloud.provider']) {
-        host['cloud.provider'] = fields['cloud.provider'];
+        pod['cloud.provider'] = fields['cloud.provider'];
       }
 
-      if (fields['cloud.instance.id']) {
-        host['cloud.instance.id'] = fields['cloud.instance.id'];
+      if (clusterName) {
+        pod['orchestrator.cluster.name'] = clusterName;
       }
 
-      if (fields['cloud.service.name']) {
-        host['cloud.service.name'] = fields['cloud.service.name'];
-      }
-
-      if (fields['cloud.region']) {
-        host['cloud.region'] = fields['cloud.region'];
-      }
-
-      if (fields['orchestrator.cluster.name']) {
-        host['orchestrator.cluster.name'] = fields['orchestrator.cluster.name'];
-      }
-
-      if (k8sPod) {
-        host['asset.children'] = [`pod:${k8sPod}`];
-      }
-
-      acc.push(host);
+      acc.push(pod);
 
       return acc;
     }, []);
